Here comes another update on eBay for the newly created screen before few days. That screen was showing the deleted Transactions & Orders that were associated with Seller.
So we have implemented new call "GetMyeBaySelling" (and for sure without this call you can't exclude the deleted orders / transactions). It was more and more of reading and exploring things in eBay area and finally we got good results - so feeling little better now :-).

Thanks Jacopo for your comment to include this feature in OFBiz trunk.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/specialpurpose/ebay@817628 13f79535-47bb-0310-9956-ffa450edef68